Add SELECT_MAP_OVERLAY and refactor SET_MAP_OVERLAY
