+++ packages/web-frontend/src/historyNavigation/historyMiddleware.js
@@ -26,12 +26,7 @@ import {
   UPDATE_MEASURE_CONFIG,
   LOCATION_CHANGE,
 } from '../actions';
-import {
-  selectCurrentPeriodGranularity,
-  selectMapOverlayById,
-  selectCurrentMapOverlayIds,
-} from '../selectors';
-import { sortMapOverlayIdsByHierarchyOrder } from '../utils';
+import { selectCurrentPeriodGranularity, selectMapOverlayByIds } from '../selectors';
 import { URL_COMPONENTS } from './constants';
 import {
   addPopStateListener,
@@ -84,24 +79,15 @@ export const historyMiddleware = store => next => action => {
       dispatchLocationUpdate(store, { [URL_COMPONENTS.REPORT_PERIOD]: null });
       break;
     case SET_MAP_OVERLAY: {
-      const mapOverlay = selectMapOverlayById(state, action.mapOverlayId);
-      if (!mapOverlay) {
+      const mapOverlays = selectMapOverlayByIds(state, action.mapOverlayIds.split(','));
+      if (mapOverlays.length === 0) {
         break;
       }
-      const { multipleMapOverlayCheckbox, mapOverlayHierarchy } = state.mapOverlayBar;
-      const currentMapOverlayIds = selectCurrentMapOverlayIds(state);
-      const mapOverlayIds = multipleMapOverlayCheckbox
-        ? sortMapOverlayIdsByHierarchyOrder(mapOverlayHierarchy, [
-            ...currentMapOverlayIds,
-            action.mapOverlayId,
-          ])
-        : [action.mapOverlayId];
 
       // TODO: PHX-1 set multiple overlay period in URL
-      const { startDate, endDate, periodGranularity } = mapOverlay;
-
+      const { startDate, endDate, periodGranularity } = mapOverlays[0];
       dispatchLocationUpdate(store, {
-        [URL_COMPONENTS.MAP_OVERLAY]: mapOverlayIds.join(','),
+        [URL_COMPONENTS.MAP_OVERLAY]: action.mapOverlayIds,
         [URL_COMPONENTS.MEASURE_PERIOD]: convertDateRangeToUrlPeriodString(
           { startDate, endDate },
           periodGranularity,
